openenv-email-triage/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-email-triage"
version = "1.0.0"
description = "OpenEnv email operations benchmark with trajectory-based rewards"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=2.7.2",
    "pydantic>=2.6,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["email_triage_env", "server"]
